Move dates.js tests to their own file
