+++ root/static/scripts/tests/index-web.js
@@ -22,4 +22,5 @@ require('./release-editor/trackParser');
 require('./release-editor/utils');
 require('./release-editor/validation');
 require('./utility');
+require('./utility/dates');
 require('./utility/sanitizedEditor');
